pyproject.toml
[project]
name = "watcher"
version = "0.1.0"
description = "A git-scraping tool with RSS feed generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "inscriptis>=2.3.0",
    "beautifulsoup4>=4.12.0",
    "feedgenerator>=2.1.0",
    "requests>=2.31.0",
    "gitpython>=3.1.0",
    "python-dateutil>=2.8.0",
    "uv",
    "ruff",
]

[project.scripts]
watcher = "watcher.cli:main"
watcher-batch = "watcher.cli_batch:main"
watcher-preview = "watcher.local_preview:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/watcher"]

[tool.uv.sources]
ask-human-mcp = { git = "https://github.com/Masony817/ask-human-mcp", rev = "24793b6362e000c12418cae7be5b8caa93cd371a" }
